+++ CHANGELOG.md
@@ -3,9 +3,13 @@ List of changes to useless-py
 
 ## [Unreleased]
 
+## [1.0.3] - 2016-07-24
+### Added
+- @nocase decorator - class decorator allowing access to attributes regardless of coding style (camelCase or snake_case)
+
 ## [1.0.2] - 2016-07-24
 ### Added
-- @extends decorator - inheritance using a decorator
+- @extends decorator - inheritance using a class decorator
 
 ## [1.0.1] - 2016-07-23
 ### Added
@@ -16,6 +20,7 @@ List of changes to useless-py
 - set_time_limit function - limits the maximum execution time of a function
 - @time_limit decorator (decorator for set_time_limit)
 
-[Unreleased]: https://github.com/Code-ReaQtor/useless-py/compare/v1.0.2...master
+[Unreleased]: https://github.com/Code-ReaQtor/useless-py/compare/v1.0.3...master
+[1.0.3]: https://github.com/Code-ReaQtor/useless-py/releases/tag/1.0.3
 [1.0.2]: https://github.com/Code-ReaQtor/useless-py/releases/tag/1.0.2
 [1.0.1]: https://github.com/Code-ReaQtor/useless-py/releases/tag/1.0.1